backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.9
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-leaflet==0.29.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
numpy==1.26.3
pandas==2.1.4
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4